VisionTransforms

Documentation for VisionTransforms.

VisionTransforms.ColorJitter

VisionTransforms.ComposedTransform

VisionTransforms.Crop

VisionTransforms.Flip

VisionTransforms.Grayscale

VisionTransforms.InvertColor

VisionTransforms.Normalize

VisionTransforms.PerImageNormalize

VisionTransforms.PerImageScale

VisionTransforms.Resize

VisionTransforms.Rot90

VisionTransforms.Scale

VisionTransforms.Solarize

VisionTransforms.Zoom

VisionTransforms.CenterCrop

VisionTransforms.CenterZoom

VisionTransforms.FlipX

VisionTransforms.FlipY

VisionTransforms.FlipZ

VisionTransforms.OriginCrop

VisionTransforms.RandomCrop

VisionTransforms.RandomZoom

VisionTransforms.add_noise

VisionTransforms.adjust_brightness

VisionTransforms.adjust_contrast

VisionTransforms.blur

VisionTransforms.color_jitter

VisionTransforms.crop

VisionTransforms.grayscale

VisionTransforms.imresize

VisionTransforms.invert_color

VisionTransforms.linear_stretch

VisionTransforms.multiply_noise

VisionTransforms.normalize

VisionTransforms.per_image_linear_stretch

VisionTransforms.per_image_normalize

VisionTransforms.permute_channels

VisionTransforms.posterize

VisionTransforms.rot90

VisionTransforms.sharpen

VisionTransforms.shift_hue

VisionTransforms.solarize

VisionTransforms.ColorJitter — Type

ColorJitter(;strength=5)

Apply a random color jitter consisting of contrast, brightness, and hue adjustments. Use strength to determine the maximum extent of the jittering, with a value in the range [1,10].

VisionTransforms.ComposedTransform — Type

ComposedTransform(transforms...)

Apply transforms to the input in the same order as they are given.

Example

julia> r = Raster(rand(256,256, 3), (X,Y,Band));

julia> t = Resample(2.0) |> Tensor();

julia> apply(t, Image(), r, 123) |> size
(512, 512, 3, 1)

julia> apply(t, Image(), r, 123) |> typeof
Array{Float32, 4}

VisionTransforms.Crop — Type

Crop(from::CropFrom, sz::Tuple)

Crop a raster of size sz from the location specified by from.

VisionTransforms.Flip — Type

Flip(dim::Int; p=1.0)

Flip the input raster along the specified dimension with probability p.

VisionTransforms.Grayscale — Type

Grayscale(;p=1.0)

Convert to grayscale by averaging the channel dimension with probability p.

VisionTransforms.InvertColor — Type

InvertColor(;p=1.0)

Apply a random color inversion with probability p.

VisionTransforms.Normalize — Type

Normalize(;mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])

Normalize the channels to have a mean of 0 and a standard deviation of 1.

Parameters

mean: The channel-wise mean of the input data (uses the ImageNet mean by default).

std: The channel-wise standard deviation of the input data (uses the ImageNet std by default).

VisionTransforms.PerImageNormalize — Type

PerImageNormalize()

Normalize the channels to have a mean of 0 and a standard deviation of 1 based on statistics calculated for each image in a batch.

VisionTransforms.PerImageScale — Type

PerImageScale(;lower=0.02, upper=0.98)

Apply a linear stretch to scale all values to the range [0, 1]. The arguments lower and upper specify the percentiles at which to define the lower and upper bounds from each channel in the source image. Values that either fall below lower or above upper will be clamped.

Parameters

lower: The quantile to use as the lower-bound in the source array.

upper: The quantile to use as the upper-bound in the source array.

VisionTransforms.Resize — Type

Resize(sz::Tuple)

Resample x according to the specified scale. Mask types will always be resampled with :near interpolation, whereas Images will be resampled with either :bilinear (scale > 1) or :average (scale < 1).

Parameters

x: The image/mask to be resampled.

sz: The size of the output image.

VisionTransforms.Rot90 — Type

Rot90(;p=1.0)

Apply a random 90 degree rotation with probability p.

VisionTransforms.Scale — Type

Scale(lower::Vector{<:Real}, upper::Vector{<:Real})

Apply a linear stretch to scale all values to the range [0, 1]. The arguments lower and upper specify the lower and upper bounds from each channel in the source image. Values that either fall below lower or above upper will be clamped.

Parameters

lower: The lower-bounds to use for each channel in the source image.

upper: The upper-bounds to use for each channel in the source image.

VisionTransforms.Solarize — Type

Solarize(;p=1.0, threshold=0.75)

Randomly solarize an image with probability p.

VisionTransforms.Zoom — Type

Zoom(from::CropFrom; strength=1.0:0.1:2.0)

Zoom to the location specified by from by a factor sampled from strength.

VisionTransforms.CenterCrop — Method

CenterCrop(sz)

Crops a tile of size sz from the center.

VisionTransforms.CenterZoom — Method

CenterZoom(;strength=1.0:0.1:2.0)

Zoom to the center of a mask/image by a factor sampled from strength.

VisionTransforms.FlipX — Method

FlipX(;p=1.0)

Apply a random horizontal flip with probability p.

VisionTransforms.FlipY — Method

FlipY(;p=1.0)

Apply a random vertical flip with probability p.

VisionTransforms.FlipZ — Method

FlipZ(;p=1.0)

Apply a random depth flip with probability p.

VisionTransforms.OriginCrop — Method

OriginCrop(sz)

Crops a tile of size sz from the origin.

VisionTransforms.RandomCrop — Method

RandomCrop(sz)

Crops a randomly placed tile of size sz.

VisionTransforms.RandomZoom — Method

RandomZoom(;strength=1.0:0.1:2.0)

Zoom to a random point by a factor sampled from strength.

VisionTransforms.add_noise — Method

add_noise(seed::Integer, dist::Distributions.Distribution, x::AbstractArray{<:Real,N}; channeldim=N, correlated=true)

Add noise generated by the distribution dist to the image tensor x.

Parameters

seed: A seed to make the outcome reproducible.

dist: A Distributions.Distribution object from which to sample the noise.

x: A tensor containing a 2D or 3D image or image series.

correlated: If true, applies the same noise value to each channel in the image.

VisionTransforms.adjust_brightness — Method

adjust_brightness(x::AbstractArray, brightness::Real)

Adjust the brightness of x by brightness.

VisionTransforms.adjust_contrast — Method

adjust_contrast(x::AbstractArray, contrast::Real, channeldim::Int)

Adjust the contrast of x by contrast.

VisionTransforms.blur — Method

blur(x::AbstractArray{<:Real}, strength::Tuple; ndims::Integer=2)
blur(x::AbstractArray{<:Real}, strength::Number; ndims::Integer=2)

Blur the image x by applying a gaussian filter with a standard deviation of strength.

VisionTransforms.color_jitter — Method

color_jitter(rng, x::AbstractArray, strength::Int, [channeldim])

Applies random color jittering transformations (hue, saturation, and brightness) to the input image x.

Parameters

rng: A random number generator to make the outcome reproducible.

x: A tensor containing an RGB image or image series.

strength: The strength of the jittering, from 1 to 10.

channeldim: The dimension corresponding to channels in x.

VisionTransforms.crop — Method

crop([rng], from::CropFrom, sz::NTuple{N1,Int}, x::AbstractArray{<:Any,N2})

Crop x to sz with the upper-left corner specified by from.

VisionTransforms.grayscale — Method

grayscale(x::AbstractArray, [channeldim])

Convert x to a grayscale image.

VisionTransforms.imresize — Method

imresize(img::AbstractArray, sz::Tuple, method::Symbol)

Resize img to sz with the specified resampling method.

Parameters

img: The image to be resized.

sz: The width and height of the output as a tuple.

method: Either :nearest or :bilinear.

VisionTransforms.invert_color — Method

invert_color(x::AbstractArray{<:Real})

Invert the colors of x.

VisionTransforms.linear_stretch — Method

linear_stretch(x::AbstractArray, lower::Vector{<:Real}, upper::Vector{<:Real}, channel_dim::Int)

Perform a linear histogram stretch on x such that lower is mapped to 0 and upper is mapped to 1. Values outside the interval [lower, upper] will be clamped.

VisionTransforms.multiply_noise — Method

multiply_noise(seed::Integer, dist::Distributions.Distribution, x::AbstractArray{<:Real,N}; channeldim=N, correlated=true)

Multiply noise generated by the distribution dist to the image tensor x.

Parameters

seed: A seed to make the outcome reproducible.

dist: A Distributions.Distribution object from which to sample the noise.

x: A tensor containing a 2D or 3D image or image series.

correlated: If true, applies the same noise value to each channel in the image.

VisionTransforms.normalize — Method

normalize(x::AbstractArray{<:Number,N}, μ::AbstractVector, σ::AbstractVector; channeldim=N)

Normalize the input array with respect to the specified dimension so that the mean is 0 and the standard deviation is 1.

Parameters

x: A tensor containing one or more 2D or 3D images or 2D image series.

μ: A Vector of means for each index in channeldim.

σ: A Vector of standard deviations for each index in channeldim.

channeldim: The dimension corresponding to channels in x.

VisionTransforms.per_image_linear_stretch — Method

per_image_linear_stretch(x::AbstractArray, lower::Real, upper::Real, channel_dim::Int)

Apply a linear stretch to scale all values to the range [0, 1]. The arguments lower and upper specify the percentiles at which to define the lower and upper bounds from each channel in the source image. Values that either fall below lower or above upper will be clamped.

VisionTransforms.per_image_normalize — Method

per_image_normalize(x::AbstractArray{<:Number,N}; channeldim=N, obsdim=nothing)

Normalize the input array so that the mean and standard deviation of each channel is 0 and 1, respectively. Unlike normalize, this method will compute new statistics for each image in x. This is more computationally expensive, but may be more suitable when there is significant domain shift between train and test images.

Parameters

x: A tensor containing one or more 2D or 3D images or 2D image series.

channeldim: The dimension corresponding to the image channel.

obsdim: The dimension corresponding to observations if present.

VisionTransforms.permute_channels — Method

permute_channels(rng, x::AbstractArray, [channeldim])

Permute the channel ordering of x.

VisionTransforms.posterize — Method

posterize(x::AbstractArray, nbits::Int)

Posterize an image x by reducing the number of bits for each color channel to nbits.

VisionTransforms.rot90 — Method

rot90(x)

Rotate the image x by 90 degress.

VisionTransforms.sharpen — Method

sharpen(x::AbstractArray, strength::Real, channeldim::Int)

Sharpen the image x by applying a high-frequency-boosting filter.

VisionTransforms.shift_hue — Method

shift_hue(x::AbstractArray, shift::Real, [channeldim])

Shift the hue of x in the HSV color space by the amount specified by shift.

VisionTransforms.solarize — Method

solarize(x::AbstractArray{<:Real}; threshold=0.75)

Solarize x by inverting all values above threshold.