VisionTransforms

Documentation for VisionTransforms.

VisionTransforms.ComposedTransform

VisionTransforms.FlipX

VisionTransforms.FlipY

VisionTransforms.RandomCrop

VisionTransforms.Resize

VisionTransforms.Rot90

VisionTransforms.apply_transform

VisionTransforms.crop

VisionTransforms.flipX

VisionTransforms.flipY

VisionTransforms.imresize

VisionTransforms.rot90

VisionTransforms.ComposedTransform — Type

ComposedTransform(transforms...)

Apply transforms to the input in the same order as they are given.

Example

julia> r = Raster(rand(256,256, 3), (X,Y,Band));

julia> t = Resample(2.0) |> Tensor();

julia> apply(t, Image(), r, 123) |> size
(512, 512, 3, 1)

julia> apply(t, Image(), r, 123) |> typeof
Array{Float32, 4}

VisionTransforms.FlipX — Type

FlipX(p)

Apply a random horizontal flip with probability p.

VisionTransforms.FlipY — Type

FlipY(p)

Apply a random vertical flip with probability p.

VisionTransforms.RandomCrop — Type

RandomCrop(size::Int)
RandomCrop(size::Tuple{Int,Int})

Crop a randomly placed tile equal to size from the input array.

VisionTransforms.Resize — Type

Resample(scale)

Resample x according to the specified scale. Mask types will always be resampled with :near interpolation, whereas Images will be resampled with either :bilinear (scale > 1) or :average (scale < 1).

Parameters

x: The raster/stack to be resampled.

scale: The size of the output with respect to the input.

VisionTransforms.Rot90 — Type

Rot90(p)

Apply a random 90 degree rotation with probability p.

VisionTransforms.apply_transform — Method

transform(t::AbstractTransform, dtype::DType, x)
transform(t::AbstractTransform, dtypes::Tuple, x::Tuple)

Apply the transformation t to the input x with data type dtype.

VisionTransforms.crop — Function

crop(x, sz::Int, ul=(1,1))
crop(x, sz::Tuple{Int,Int}, ul=(1,1))

Crop a tile equal to size out of x with an upper-left corner defined by ul.

VisionTransforms.flipX — Method

flipX(x)

Flip the image x across the horizontal axis.

VisionTransforms.flipY — Method

flipY(x)

Flip the image x across the vertical axis.

VisionTransforms.imresize — Method

resize(img::AbstractArray, sz::Tuple{Int,Int}; method=:bilinear)

Resize img to sz with the specified resampling method.

Parameters

img: The image to be resized.

sz: The width and height of the output as a tuple.

method: Either :nearest or :bilinear.

VisionTransforms.rot90 — Method

rot90(x)

Rotate the image x by 90 degress.