VisionTransforms

Documentation for VisionTransforms.

VisionTransforms.ColorJitter

VisionTransforms.ComposedTransform

VisionTransforms.FlipX

VisionTransforms.FlipY

VisionTransforms.Normalize

VisionTransforms.PerImageNormalize

VisionTransforms.PerImageScale

VisionTransforms.RandomCrop

VisionTransforms.Resize

VisionTransforms.Rot90

VisionTransforms.Scale

VisionTransforms.add_noise

VisionTransforms.color_jitter

VisionTransforms.crop

VisionTransforms.flipX

VisionTransforms.flipY

VisionTransforms.imresize

VisionTransforms.linear_stretch

VisionTransforms.multiply_noise

VisionTransforms.normalize

VisionTransforms.per_image_linear_stretch

VisionTransforms.per_image_normalize

VisionTransforms.random_crop

VisionTransforms.rot90

VisionTransforms.transform

VisionTransforms.ColorJitter — Type

ColorJitter(;contrast=0.8:0.1:1.2, brightness=-0.2:0.1:0.2)

Apply a random color jittering transformations (contrast and brightness adjustments) according to the formula α * x + β * M, where α is contrast, β is brightness, and M is either the mean or maximum value of x.

VisionTransforms.ComposedTransform — Type

ComposedTransform(transforms...)

Apply transforms to the input in the same order as they are given.

Example

julia> r = Raster(rand(256,256, 3), (X,Y,Band));

julia> t = Resample(2.0) |> Tensor();

julia> apply(t, Image(), r, 123) |> size
(512, 512, 3, 1)

julia> apply(t, Image(), r, 123) |> typeof
Array{Float32, 4}

VisionTransforms.FlipX — Type

FlipX(p)

Apply a random horizontal flip with probability p.

VisionTransforms.FlipY — Type

FlipY(p)

Apply a random vertical flip with probability p.

VisionTransforms.Normalize — Type

Normalize(;mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])

Normalize the channels to have a mean of 0 and a standard deviation of 1.

Parameters

mean: The channel-wise mean of the input data (uses the ImageNet mean by default).

std: The channel-wise standard deviation of the input data (uses the ImageNet std by default).

VisionTransforms.PerImageNormalize — Type

PerImageNormalize()

Normalize the channels to have a mean of 0 and a standard deviation of 1 based on statistics calculated for each image in a batch.

VisionTransforms.PerImageScale — Type

PerImageScale(;lower=0.02, upper=0.98)

Apply a linear stretch to scale all values to the range [0, 1]. The arguments lower and upper specify the percentiles at which to define the lower and upper bounds from each channel in the source image. Values that either fall below lower or above upper will be clamped.

Parameters

lower: The quantile to use as the lower-bound in the source array.

upper: The quantile to use as the upper-bound in the source array.

VisionTransforms.RandomCrop — Type

RandomCrop(size::Int)
RandomCrop(size::Tuple{Int,Int})

Crop a randomly placed tile equal to size from the input array.

VisionTransforms.Resize — Type

Resize(sz::Tuple)

Resample x according to the specified scale. Mask types will always be resampled with :near interpolation, whereas Images will be resampled with either :bilinear (scale > 1) or :average (scale < 1).

Parameters

x: The image/mask to be resampled.

sz: The size of the output image.

VisionTransforms.Rot90 — Type

Rot90(p)

Apply a random 90 degree rotation with probability p.

VisionTransforms.Scale — Type

Scale(lower::Vector{<:Real}, upper::Vector{<:Real})

Apply a linear stretch to scale all values to the range [0, 1]. The arguments lower and upper specify the lower and upper bounds from each channel in the source image. Values that either fall below lower or above upper will be clamped.

Parameters

lower: The lower-bounds to use for each channel in the source image.

upper: The upper-bounds to use for each channel in the source image.

VisionTransforms.add_noise — Method

add_noise([rng], dist::Distributions.Distribution, x::AbstractArray; correlated=true)

Add noise generated by the distribution dist to the image tensor x.

Parameters

rng: A random number generator to make the outcome reproducible.

dist: A Distributions.Distribution object from which to sample the noise.

x: A tensor containing a 2D or 3D image or image series.

correlated: If true, applies the same noise value to each channel in the image.

VisionTransforms.color_jitter — Method

color_jitter([rng], x::Image2D, contrast, brightness; kw...)
color_jitter([rng], x::Image3D, contrast, brightness; kw...)
color_jitter([rng], x::Series2D, contrast, brightness; kw...)
color_jitter([rng], x::AbstractArray, contrast, brightness, dims; usemax=true)

Applies random color jittering transformations (contrast and brightness adjustments) to input images or data series according to the formula α * x + β * M, where α is contrast, β is brightness, and M is either the mean or maximum value of x.

Parameters

rng: A random number generator to make the outcome reproducible.

dist: A Distributions.Distribution object from which to sample the noise.

x: A tensor containing a 2D or 3D image or image series.

correlated: If true, applies the same noise value to each channel in the image.

VisionTransforms.crop — Function

crop(x, sz::Int, ul=(1,1))
crop(x, sz::Tuple{Int,Int}, ul=(1,1))

Crop a tile equal to size out of x with an upper-left corner defined by ul.

VisionTransforms.flipX — Method

flipX(x)

Flip the image x across the horizontal axis.

VisionTransforms.flipY — Method

flipY(x)

Flip the image x across the vertical axis.

VisionTransforms.imresize — Method

imresize(img::AbstractArray, sz::Tuple{Int,Int}; method=:bilinear)

Resize img to sz with the specified resampling method.

Parameters

img: The image to be resized.

sz: The width and height of the output as a tuple.

method: Either :nearest or :bilinear.

VisionTransforms.linear_stretch — Method

linear_stretch(x::Image2D, lower::Vector{<:Real}, upper::Vector{<:Real})
linear_stretch(x::Image3D, lower::Vector{<:Real}, upper::Vector{<:Real})
linear_stretch(x::Series2D, lower::Vector{<:Real}, upper::Vector{<:Real})
linear_stretch(x::AbstractArray, lower::Vector{<:Real}, upper::Vector{<:Real}, channel_dim::Int)

Perform a linear histogram stretch on x such that lower is mapped to 0 and upper is mapped to 1. Values outside the interval [lower, upper] will be clamped.

VisionTransforms.multiply_noise — Method

multiply_noise([rng], dist::Distributions.Distribution, x::AbstractArray; correlated=true)

Multiply noise generated by the distribution dist to the image tensor x.

Parameters

rng: A random number generator to make the outcome reproducible.

dist: A Distributions.Distribution object from which to sample the noise.

x: A tensor containing a 2D or 3D image or image series.

correlated: If true, applies the same noise value to each channel in the image.

VisionTransforms.normalize — Method

normalize(x::AbstractArray, μ::AbstractVector, σ::AbstractVector; dim=1)

Normalize the input array with respect to the specified dimension so that the mean is 0 and the standard deviation is 1.

Parameters

μ: A Vector of means for each index in dim.

σ: A Vector of standard deviations for each index in dim.

dim: The dimension along which to normalize the input array.

VisionTransforms.per_image_linear_stretch — Method

per_image_linear_stretch(x::Image2D, lower::Real, upper::Real)
per_image_linear_stretch(x::Image3D, lower::Real, upper::Real)
per_image_linear_stretch(x::Series2D, lower::Real, upper::Real)
per_image_linear_stretch(x::AbstractArray, lower::Real, upper::Real, channel_dim::Int)

Apply a linear stretch to scale all values to the range [0, 1]. The arguments lower and upper specify the percentiles at which to define the lower and upper bounds from each channel in the source image. Values that either fall below lower or above upper will be clamped.

VisionTransforms.per_image_normalize — Method

per_image_normalize(x::Image2D)
per_image_normalize(x::Image3D)
per_image_normalize(x::Series2D)
per_image_normalize(x::AbstractArray, dims::Tuple)

Normalize the input array so that the mean and standard deviation of each channel is 0 and 1, respectively. Unlike normalize, this method will compute new statistics for each image in x. This is more computationally expensive, but may be more suitable when there is significant domain shift between train and test images.

Parameters

x: A tensor containing one or more 2D or 3D images or 2D image series.

dims: The dimensions over which to compute image statistics.

VisionTransforms.random_crop — Method

random_crop(seed::Int, x::AbstractArray, sz::Tuple{Int,Int})

Crop a randomly placed tile equal to sz from the array x.

VisionTransforms.rot90 — Method

rot90(x)

Rotate the image x by 90 degress.

VisionTransforms.transform — Method

transform(t::AbstractTransform, dtype::DType, x)
transform(t::AbstractTransform, dtypes::Tuple, x::Tuple)

Apply the transformation t to the input x with data type dtype.